Create README.md
Created a README with a simple description of what this project is, and the technologies involved in creating it. 

So far, I've only described the purpose of the project, and the two primary fields of AI that are relevant to the field. I still need to add implementation details, but that will come as I build the project.